requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.2.1=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py310h9128911_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.5.7=h56e8100_0
certifi=2022.12.7=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310haa95532_0
colorama=0.4.6=py310haa95532_0
contourpy=1.0.5=py310h59b6b97_0
cryptography=39.0.1=py310h21b164f_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fftw=3.3.9=h2bbff1b_1
flask=2.2.2=py310haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h546665d_1
future=0.18.3=py310haa95532_0
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.0=pyh025b116_0
ipython=8.13.2=pyh08f2357_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py310haa95532_0
joblib=1.1.1=py310haa95532_0
jpeg=9e=h8ffe710_2
jsonlines=3.1.0=pyhd8ed1ab_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py310h5588dad_0
kiwisolver=1.4.4=py310h476a331_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.17=h8ffe710_0
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
lz4-c=1.9.3=h8ffe710_1
markupsafe=2.1.1=py310h2bbff1b_0
matplotlib=3.7.1=py310h5588dad_0
matplotlib-base=3.7.1=py310h4ed8f06_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
munkres=1.1.4=pyh9f0ad1d_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numexpr=2.8.4=py310hd213c9f_0
numpy=1.24.3=py310hdc03b94_0
numpy-base=1.24.3=py310h3caf3d7_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.2=py310h4ed8f06_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310hd77b12b_0
pip=23.0.1=py310haa95532_0
platformdirs=3.5.0=pyhd8ed1ab_0
ply=3.11=py_1
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.0.0=py310haa95532_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py310hd77b12b_0
pyqt5-sip=12.11.0=py310hd77b12b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.11=h966fe2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.12.1=cpu_py310h5e1f01c_1
pytz=2022.7=py310haa95532_0
pywin32=305=py310h2bbff1b_0
pyyaml=6.0=py310h2bbff1b_1
pyzmq=25.0.2=py310hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h2bbfb41_5
requests=2.29.0=py310haa95532_0
scikit-learn=1.2.0=py310hd77b12b_1
scipy=1.9.3=py310h86744a3_0
setuptools=66.0.0=py310haa95532_0
sip=6.6.2=py310h8a704f9_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd8ed1ab_0
torchvision=0.13.1=cpu_py310h378ed51_0
tornado=6.2=py310he2412df_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.5.0=py310haa95532_0
typing_extensions=4.5.0=py310haa95532_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.2.3=py310haa95532_0
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
